net: bpf: fix return type mismatch in bpf_sk_fullsock for Clang
When building kernel 4.9 with Clang, the following error occurs:
"incompatible pointer to integer conversion returning 'void *'
from a function with result type 'u64'"

The function bpf_sk_fullsock is defined to return a u64 type,
but the code currently returns the NULL macro, which is a pointer
(void *). GCC often ignores this implicit conversion with a warning,
but Clang strictly prohibits it and treats it as an error.

Cast NULL to unsigned long to satisfy Clang's strict type checking
and allow the build to proceed.

+++ net/core/filter.c
@@ -1536,7 +1536,7 @@ BPF_CALL_1(bpf_sk_fullsock, struct sock *, sk)
 {
 //	sk = sk_to_full_sk(sk);
 
-	return NULL; //sk_fullsock(sk) ? (unsigned long)sk : (unsigned long)NULL;
+	return (unsigned long)NULL; //sk_fullsock(sk) ? (unsigned long)sk : (unsigned long)NULL;
 }
 
 static const struct bpf_func_proto bpf_sk_fullsock_proto = {
